backend: simple in-memory store implementation
This is just for preliminar testing and subject to be completely
changed, so tests are not yet available. Once schema stabilizes and we
have a better idea of what we need we'd start doing things right.

The change includes Invoice and Customer mutations and queries. It
also adds some basic test data that may be useful for
prototyping. Here are some example workable test queries:

```
query invoices {
  invoices {
    identifier
    date
    from {
      name
    }
    to {
      name
    }
    items {
      description
      quantity
      unitPrice
    }
  }
}

mutation addInvoice {
  createInvoice(invoice: {
    identifier: "#0002",
    organizationId: "1",
    customerId: "1",
    date: "2019-05-27T15:09:17-03:00"
    items: [
      {
        description: "decription",
        quantity:"1.0",
        unitPrice:"500.0"
      }
    ]
  }) {
    identifier
  }
}
```